refactor(tests): remove unused test setup files and update helpers export
