--- packages/fast-element/src/__test__/setup-browser.cjs
@@ -1,6 +1,0 @@
-function importAll(r) {
-    r.keys().forEach(r);
-}
-
-// Explicitly add to browser test
-importAll(require.context("../../dist/esm/", true, /\.spec\.js$/));
