+++ packages/fast-router/.npmignore
@@ -1,6 +1,4 @@
 # Tests
-dist/dts/__test__/
-dist/esm/__test__/
 *.spec.*
 coverage/
